Outsourced reply logic to crate::replies::reply_to
