+++ src/config.rs
@@ -65,3 +65,8 @@ impl Config {
         out
     }
 }
+
+pub enum Bots {
+    Autokommentator,
+    KaenguruKnecht,
+}
